+++ test/helper/proxy.ts
@@ -15,20 +15,9 @@ export async function setupProxy(node: ChainNode) {
   await node.clearAllTransactions()
 
   // do we need proxys for other personas?
-  // Eve delegating to Bob
-  // const extrinsicEve = await node.addProxy(
-  //   '//Eve', // e.g. //Alice
-  //   '5FHneW46xGXgs5mUiveU4sbTyGBzmstUspZC92UhjJM694ty', //e.g. '5DAAnrj7VHTznn2AWBemMuyBwZWs6FNFjdyVXUeYum3PTXFy'
-  //   'RunProcess',
-  //   0
-  // )
-  // await node.submitRunProcessForProxy(extrinsicEve)
-
-  // await node.sealBlock()
 }
 
 export async function removeProxy(node: ChainNode) {
-  // Alice is Proxy for Dave == Dave is delegating to Alice
   const extrinsicDave = await node.removeProxy(proxyReq)
   await node.submitRunProcessForProxy(extrinsicDave)
 
@@ -37,13 +26,10 @@ export async function removeProxy(node: ChainNode) {
 
 export async function withProxy(node: ChainNode) {
   before(async function () {
-    // does this throw an error?
-    // apparently mocha will skip after hooks if the before hook fails
     await removeProxy(node)
     await setupProxy(node)
   })
   after(async function () {
     await removeProxy(node)
-    console.log('in after')
   })
 }
