+++ test/integration/onchain/demandB.test.ts
@@ -7,17 +7,21 @@ import { post } from '../../helper/routeHelper.js'
 import { seed, cleanup, seededDemandBId, parametersAttachmentId } from '../../seeds/onchainSeeds/demandB.seed.js'
 import { selfAddress, withIdentitySelfMock } from '../../helper/mock.js'
 import Database, { DemandRow } from '../../../src/lib/db/index.js'
-import ChainNode from '../../../src/lib/chainNode.js'
 import { pollDemandCommentState, pollDemandState, pollTransactionState } from '../../helper/poll.js'
 import { withAppAndIndexer } from '../../helper/chainTest.js'
 import { container } from 'tsyringe'
 import { filterRejectedAndAcceptedPromises } from '../../helper/parallelTests.js'
 import { withProxy } from '../../helper/proxy.js'
+import ExtendedChainNode from '../../helper/testInstanceChainNode.js'
+import { logger } from '../../../src/lib/logger.js'
+import env from '../../../src/env.js'
 
-describe.only('on-chain', function () {
+describe('on-chain', function () {
   this.timeout(60000)
   const db = new Database()
-  const node = container.resolve(ChainNode)
+  container.registerInstance(ExtendedChainNode, new ExtendedChainNode(logger, env))
+
+  const node = container.resolve(ExtendedChainNode)
   const context: { app: Express; indexer: Indexer } = {} as { app: Express; indexer: Indexer }
 
   withAppAndIndexer(context)
